Rename internal RationalLinearSubspace to RationalAffineSubspace
