+++ libintervalxt/benchmark/Makefile.am
@@ -1,6 +1,6 @@
 noinst_PROGRAMS = benchmark
 
-benchmark_SOURCES = main.cc induction.benchmark.cc rational_linear_subspace.benchmark.cc ../src/rational_linear_subspace.cc
+benchmark_SOURCES = main.cc induction.benchmark.cc rational_affine_subspace.benchmark.cc ../src/rational_affine_subspace.cc
 
 AM_CPPFLAGS = -I $(srcdir)/.. -I $(builddir)/..
 AM_LDFLAGS = $(builddir)/../src/libintervalxt.la
