+++ libintervalxt/src/impl/rational_affine_subspace.hpp
